refac: tobe 메뉴 테스트 - 하위 패키지 분류
