+++ src/test/java/kitchenpos/menus/tobe/domain/menu/MenuNameTest.java
@@ -1,9 +1,8 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.menu;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
-import kitchenpos.menus.tobe.domain.menu.MenuName;
 import kitchenpos.mock.client.FakePurgomalumClient;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
