+++ src/test/java/kitchenpos/menus/tobe/domain/menugroup/MenuGroupIdTest.java
@@ -1,37 +1,36 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.menugroup;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
 import java.util.UUID;
-import kitchenpos.menus.tobe.domain.menu.MenuId;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullSource;
 
-class MenuIdTest {
+class MenuGroupIdTest {
 
-    @DisplayName("UUID로 메뉴 ID를 생성할 수 있다.")
+    @DisplayName("UUID로 메뉴 그룹 ID를 생성할 수 있다.")
     @Test
     void create() {
         // given
         UUID id = UUID.randomUUID();
 
         // when
-        MenuId menuId = new MenuId(id);
+        MenuGroupId menuGroupId = new MenuGroupId(id);
 
         // then
-        assertThat(menuId.getValue()).isEqualTo(id);
+        assertThat(menuGroupId.getValue()).isEqualTo(id);
     }
 
-    @DisplayName("메뉴 ID 값이 올바르지 않으면, 예외가 발생한다.")
+    @DisplayName("메뉴 그룹 ID 값이 올바르지 않으면, 예외가 발생한다.")
     @ParameterizedTest
     @NullSource
-    void createWithNull(UUID value) {
+    void createWitNull(UUID value) {
         // when then
-        assertThatThrownBy(() -> new MenuId(value))
+        assertThatThrownBy(() -> new MenuGroupId(value))
             .isInstanceOf(IllegalArgumentException.class)
-            .hasMessage(MenuId.ERROR_MESSAGE_VALUE_NULL);
+            .hasMessage(MenuGroupId.ERROR_MESSAGE_VALUE_NULL);
     }
 }
