+++ src/test/java/kitchenpos/menus/tobe/domain/menugroup/MenuGroupTest.java
@@ -1,9 +1,8 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.menugroup;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.UUID;
-import kitchenpos.menus.tobe.domain.menugroup.MenuGroup;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
