+++ src/test/java/kitchenpos/menus/tobe/domain/menuproduct/MenuProductsTest.java
@@ -1,12 +1,9 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.menuproduct;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.math.BigDecimal;
 import java.util.List;
-import kitchenpos.menus.tobe.domain.menuproduct.MenuProduct;
-import kitchenpos.menus.tobe.domain.menuproduct.MenuProductQuantity;
-import kitchenpos.menus.tobe.domain.menuproduct.MenuProducts;
 import kitchenpos.mock.fixture.ProductFixture;
 import kitchenpos.products.tobe.domain.Product;
 import org.junit.jupiter.api.DisplayName;
